// types/mapping.ts
export interface MappedData {
    runnerName: string;
    matchId: string;
    questionId: number;
    optionId: number;
    optionName: string;
    marketId: string;
    eventId: string;
    selectionId: string;
    type: 'fancy-odds' | 'bookmaker-odds' | 'match-odds';
}

export interface PredictionRequest {
    market_id: string;
    event_id: string;
    invest_amount: number;
    RunnerName: string;
    match_id: string;
    betquestion_id: number;
    betoption_id: number;
    Option_name: string;
    ratio: string;
    selection_id: string;
    type: string;
    is_back: boolean;
    level: number;
}